refactored CreatePoolSteps Component Code
- moved logic for each CreatePool step into its own custom hook
- removed unused state
- moved styling into its own file
- commented on code that can be ambiguous
- drastically reduced file size in CreatePoolSteps by doing the methods above
