+++ swimster-ui/src/components/CreatePoolSteps/style.js
@@ -1,0 +1,26 @@
+export const style = {
+    formCreatePoolHeader: 'font-semibold text-3xl col-span-3 mb-3',
+    formImagesUploadHeader: 'col-span-3',
+    formImagesUploadTitle: 'font-semibold text-2xl',
+    formImagesUploadSubtitle: 'font-semibold text-sm text-gray-500',
+    imageUploadButton: 'hidden',
+    imageUploadLabel: 'bg-blue-700 text-gray-200 font-semibold py-2 text-center rounded-md cursor-pointer col-span-3',
+    renderSelectedImagesContainer: 'col-span-3',
+    formImagesUploadButtons: 'col-span-3 flex justify-end space-x-4',
+    formImagesUploadBackButton: 'bg-blue-700 p-3 w-1/4 rounded-md cursor-pointer font-semibold text-gray-200',
+    formImagesUploadContinueButton: 'bg-blue-700 p-3 w-1/4 rounded-md cursor-pointer font-semibold text-gray-200',
+    formAmenitiesHeader: 'font-semibold text-2xl col-span-3',
+    formDimensionsHeader: 'font-semibold text-2xl col-span-3',
+    amenitiesContainer: 'col-span-3',
+    amenitiesContent: 'grid grid-cols-3 gap-6 mt-3',
+    amenityContainer: 'rounded-xl flex flex-col space-y-2 items-center border px-3 py-4 font-semibold cursor-pointer',
+    selectedAmenity: 'bg-gray-500 text-gray-200 border-gray-500',
+    checked: 'bg-blue-100',
+    amenityTitle: 'text-sm',
+    amenityImage: 'w-8 h-8 fill-gray-200',
+    inputElement: 'rounded-md p-3 border-none outline-none bg-gray-200 ring-gray-400 hover:ring-2 focus:ring-4 focus:ring-offset-2 focus:ring-offset-gray-300',
+    fullInput: 'col-span-3 resize-none',
+    formStepContinueButton: 'bg-blue-700 py-3 col-start-3 rounded-md mt-4 cursor-pointer font-semibold text-gray-200',
+    formStepBackButton: 'bg-blue-700 py-3 col-start-2 col-span-1 rounded-md mt-4 cursor-pointer font-semibold text-gray-200',
+    formStepConfirmButton: 'bg-blue-700 py-3 col-start-3 col-span-1 rounded-md mt-4 cursor-pointer font-semibold text-gray-200',
+}
